удалил лишние файлы и почистил код

+++ forma/modules/core/controllers/UserController.php
@@ -95,8 +95,10 @@ class UserController extends Controller
         $modelUpload = $model->userProfile;
         $modelUser = UserProfile::find()->where(['user_id'=>$model->id])->one();
         if (Yii::$app->request->isPost) {
-            if(!empty(UploadedFile::getInstance($modelUpload, 'imageFile'))){
-                unlink('./img/user-profile/' . $model->userProfile->image);
+            if (!empty(UploadedFile::getInstance($modelUpload, 'imageFile'))) {
+                if (isset($model->userProfile->image)) {
+                    unlink('./img/user-profile/' . $model->userProfile->image);
+                }
                 $modelUpload->imageFile = UploadedFile::getInstance($modelUpload, 'imageFile');
                 $fileName = $modelUpload->imageFile->baseName . '_' . time() . date('Y-m-d') . '.' . $modelUpload->imageFile->extension;
                 if ($modelUpload->imageFile
@@ -104,7 +106,6 @@ class UserController extends Controller
                     $modelUser->image = $fileName;
                 }
             }
-
         }
 
         if ($model->load(Yii::$app->request->post()) && $model->save() && $modelUser->save()) {
